% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cPASS.R
\name{cPASS}
\alias{cPASS}
\title{cPASS}
\usage{
cPASS(datInput)
}
\arguments{
\item{datInput}{A dataframe with rows containing proteins and column names: Experiment.id, Replicate, Bait,
Prey, and count (i.e., prey count).}
}
\value{
Data frame containing bait-prey pairs with average peptide spectrum match (PSMs), total PSMs,
ratio total PSMs,Z-score,S-score,D-score and WD-score.
}
\description{
Apply compass to score instances (e.g., bait-prey interactons (BPIs) in the data.frame.
}
\references{
Huttlin, E. L., Ting, L., Bruckner, R. J., Gebreab, F., Gygi, M. P., Szpyt, J., et al. (2015).
The BioPlex Network: A Systematic Exploration of the Human Interactome. Cell 162, 425-440.
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
