% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diceCoefficient.R
\name{diceCoefficient}
\alias{diceCoefficient}
\title{diceCoefficient}
\usage{
diceCoefficient(datInput)
}
\arguments{
\item{datInput}{Data frame containing instances (e.g., bait-prey interactons (BPIs))}
}
\value{
Data frame containing a classification result for all instances in the dataset
}
\description{
Compute the Dice similarity coefficient scores for instances (e.g., bait-prey interactons (BPIs)) in the data.frame.
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
