% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccardCoefficient.R
\name{jaccardCoefficient}
\alias{jaccardCoefficient}
\title{jaccardCoefficient}
\usage{
jaccardCoefficient(datInput)
}
\arguments{
\item{datInput}{A dataframe with rows containing proteins and column names: Experiment.id, Replicate, Bait,
Prey, and Spectral count.}
}
\value{
Data frame containing bait-prey pairs with the jaccard coefficient, a number between 0 and 1
}
\description{
Compute the jaccard similarity coefficient scores for instances (e.g., bait-prey interactons (BPIs)) in the data.frame.
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
