% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDatInput.R
\docType{data}
\name{SampleDatInput}
\alias{SampleDatInput}
\title{Test data for scoring}
\usage{
data(SampleDatInput)
}
\description{
Bait-Prey Interactions (BPIs)
}
\details{
\itemize{
  \item Experiment ID
  \item Replicate
  \item Bait
  \item Prey
  \item counts
}
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
\keyword{datasets}
