% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpsonCoefficient.R
\name{simpsonCoefficient}
\alias{simpsonCoefficient}
\title{simpsonCoefficient}
\usage{
simpsonCoefficient(datInput)
}
\arguments{
\item{datInput}{Data frame with column names: Experiment.id, Replicate, Bait,
Prey, and count (i.e., prey count).}
}
\value{
Data frame containing bait-prey pairs with Simpson coefficient,
a number between 0 and 1
}
\description{
This function computes the Simpson similarity coefficient scores for
instances (e.g., bait-prey interactions (BPIs)) in the data.frame.
}
\examples{
data(SampleDatInput)
datScoring <- overlapScore(SampleDatInput)
head(datScoring)
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
