% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdfClassifier.R
\docType{data}
\name{testdfClassifier}
\alias{testdfClassifier}
\title{Test data for classifier}
\usage{
data(testdfClassifier)
}
\description{
Scored data.frame
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
\keyword{datasets}
