% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccardCoefficient.R
\name{jaccardCoefficient}
\alias{jaccardCoefficient}
\title{jaccardCoefficient}
\usage{
jaccardCoefficient(datInput)
}
\arguments{
\item{datInput}{Data frame with column names: Experiment.id, Replicate, Bait,
Prey, and count (i.e., prey count).}
}
\value{
Data frame containing bait-prey pairs with the Jaccard coefficient
score, a number between 0 and 1
}
\description{
This function computes the Jaccard similarity coefficient scores for instances
(e.g., bait-prey interactions (BPIs)) in the data.frame.
}
\examples{
data(SampleDatInput)
datScoring <- jaccardCoefficient(SampleDatInput)
head(datScoring)
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
