% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmTrain.R
\name{svmTrain}
\alias{svmTrain}
\title{svmTrain}
\usage{
svmTrain(
  dtInput,
  impute = TRUE,
  p = 0.3,
  parameterTuning = TRUE,
  cost = seq(from = 2, to = 10, by = 2),
  gamma = seq(from = 0.01, to = 0.1, by = 0.02),
  kernel = "radial",
  ncross = 10,
  pr.plot = TRUE,
  roc.plot = TRUE
)
}
\arguments{
\item{dtInput}{Data frame containing instances with class labels}

\item{impute}{Logical value, indicating whether to impute missing values}

\item{p}{The percentage of data that goes to training; defaults to 0.3}

\item{parameterTuning}{Logical value; indicating whether to
tune SVM hyperparameters}

\item{cost}{Cost of constraints violation}

\item{gamma}{Paramter needed for all kernels except linear}

\item{kernel}{Kernel type: 'linear', 'polynomial', 'sigmoid', or 'radial';
defaults to 'radial'}

\item{ncross}{K-fold cross validation on the training data is performed to assess the quality of the model;
defaults to 10}

\item{pr.plot}{Logcal value, indicating whether to plot precision-recall (PR) curve}

\item{roc.plot}{Logcal value, indicating whether to plot ROC curve}
}
\value{
Data frame containing a classification result for all instances in the dataset
}
\description{
Classification using the Support Vector Machines (SVM) algorithm.
The function also computes the areas under the precision-recall (PR) and ROC curve to evalute the performance of the classifer.
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
