% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weighted.matrixModel.R
\name{Weighted.matrixModel}
\alias{Weighted.matrixModel}
\title{Weighted.matrixModel}
\usage{
Weighted.matrixModel(datInput)
}
\arguments{
\item{datInput}{A dataframe with rows containing proteins and column names: Experiment.id, Replicate, Bait,
Prey, and Spectral count.}
}
\value{
Data frame containing bait-prey pairs with k (i.e.,number of co-purifications) &
logHG (i.e., $-1$*log(P-val) of the hypergeometric test)
}
\description{
Compute the weighted matrix model for instances (e.g., bait-prey interactons (BPIs)) in the data.frame.
}
\references{
Drew, K., Lee, C., Huizar, R. L., Tu, F., Borgeson, B., McWhite, C. D., et al. (2017).
Integration of over 9,000 mass spectrometry experiments builds a global map of human protein complexes.
Mol. Syst. Biol. 13, 932.
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
